BLD / MAINT: Improvements related to Windows compatibility
* adjustments to the build process, code base, and test suite
that improve Windows compatibility; many of these are code
modernizations (i.e., broader adoption of pytest) that serve
the dual purpose of improving compatibility

+++ package/MDAnalysis/lib/formats/libdcd.pyx
@@ -653,7 +653,7 @@ cdef class DCDFile:
 
         cdef np.ndarray[np.int64_t, ndim=1] c_indices
         if indices is None:
-            c_indices = np.arange(self.natoms)
+            c_indices = np.arange(self.natoms, dtype=np.int64)
             natoms = self.natoms
         else:
             natoms = len(indices)
